Changes to arkadius's comments
