+++ utils/avro-components-utils/src/main/scala/pl/touk/nussknacker/engine/avro/schemaregistry/confluent/ConfluentUtils.scala
@@ -4,10 +4,9 @@ import cats.data.Validated
 import com.typesafe.scalalogging.LazyLogging
 import io.confluent.kafka.schemaregistry.ParsedSchema
 import io.confluent.kafka.schemaregistry.avro.{AvroSchema, AvroSchemaProvider, AvroSchemaUtils}
-import io.confluent.kafka.serializers.NonRecordContainer
 import org.apache.avro.Schema
-import org.apache.avro.generic.{GenericContainer, GenericData, GenericDatumWriter, GenericRecord}
-import org.apache.avro.io.{DecoderFactory, Encoder, EncoderFactory}
+import org.apache.avro.generic.{GenericContainer, GenericDatumWriter}
+import org.apache.avro.io.{DecoderFactory, EncoderFactory}
 import org.apache.avro.specific.{SpecificDatumWriter, SpecificRecord}
 import org.apache.kafka.common.errors.SerializationException
 import pl.touk.nussknacker.engine.avro.AvroUtils
@@ -16,6 +15,7 @@ import pl.touk.nussknacker.engine.avro.schema.StringForcingDatumReaderProvider
 import java.io.{ByteArrayOutputStream, DataOutputStream, OutputStream}
 import java.nio.ByteBuffer
 import java.util
+import collection.JavaConverters._
 
 object ConfluentUtils extends LazyLogging {
 
@@ -76,7 +76,7 @@ object ConfluentUtils extends LazyLogging {
       case v: Array[Byte] =>
         output.write(v)
       case v =>
-        val schema = AvroUtils.getSchema(v)
+        val schema = getSchema(v)
 
         val writer = data match {
           case _: SpecificRecord =>
@@ -93,6 +93,29 @@ object ConfluentUtils extends LazyLogging {
     val bytes = output.toByteArray
     output.close()
     bytes
+  }
+
+  /**
+    * Discovering AvoSchema based on data
+    */
+  def getSchema(data: Any): Schema = {
+    def discoverSchema(data: List[Any]) = data.map(getSchema).distinct match {
+      case head :: Nil => head
+      case list => Schema.createUnion(list.asJava)
+    }
+
+    data match {
+      case container: GenericContainer =>
+        container.getSchema
+      case map: java.util.Map[_, _] =>
+        val mapValuesSchema = discoverSchema(map.values.asScala.toList)
+        Schema.createMap(mapValuesSchema)
+      case list: java.util.List[_] =>
+        val listValuesSchema = discoverSchema(list.asScala.toList)
+        Schema.createArray(listValuesSchema)
+      case _ =>
+        AvroSchemaUtils.getSchema(data)
+    }
   }
 
   private def writeSchemaId(schemaId: Int, stream: OutputStream): Unit = {
